requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2018.1.18
click==6.7
Flask==0.12.2
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.10
kombu==4.1.0
MarkupSafe==1.0
mysql-connector-python-rf==2.2.2
PyMySQL==0.8.0
python-dotenv==0.8.2
pytz==2018.3
redis==2.10.6
SQLAlchemy==1.2.5
vine==1.1.4
Werkzeug==0.14.1
WTForms==2.1
